# Engineer Competency/app/docs/swagger.json
{"schemes":["http"],"swagger":"2.0","info":{"description":"API Gateway Internal Auth Service dengan OAuth, wallet, dan invoice.","title":"Engineer Competency API","contact":{},"version":"1.0"},"host":"localhost:8080","basePath":"/","paths":{"/api/v1/invoice/all":{"get":{"security":[{"BearerAuth":[]}],"description":"Get list of all invoices","consumes":["application/json"],"produces":["application/json"],"tags":["Invoice"],"summary":"Get all invoices","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/invoice/create":{"post":{"security":[{"BearerAuth":[]}],"description":"Create a new invoice for a merchant","consumes":["application/json"],"produces":["application/json"],"tags":["Invoice"],"summary":"Create invoice","parameters":[{"description":"Invoice create request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/models.InvoiceRequest"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/invoice/update-status":{"put":{"security":[{"BearerAuth":[]}],"description":"Update invoice fields such as status and payment type","consumes":["application/json"],"produces":["application/json"],"tags":["Invoice"],"summary":"Update invoice status","parameters":[{"description":"Invoice update request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/models.InvoiceRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/model.Response"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/invoice/{id}":{"get":{"security":[{"BearerAuth":[]}],"description":"Get invoice details by invoice ID","consumes":["application/json"],"produces":["application/json"],"tags":["Invoice"],"summary":"Get invoice by ID","parameters":[{"type":"integer","description":"Invoice ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/oauth/token":{"post":{"security":[{"BasicAuth":[]}],"description":"Generate OAuth access token using client credentials and password","consumes":["application/json"],"produces":["application/json"],"tags":["OAuth"],"summary":"Generate OAuth token","parameters":[{"description":"OAuth token request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/models.Request"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/oauth/verify-login-token":{"post":{"security":[{"BasicAuth":[]}],"description":"Verify and validate an existing login token","consumes":["application/json"],"produces":["application/json"],"tags":["OAuth"],"summary":"Verify login token","parameters":[{"description":"Login token request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/models.Request"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/wallet/balance/{user_id}":{"get":{"security":[{"BearerAuth":[]}],"description":"Get balance for a user wallet by user ID","consumes":["application/json"],"produces":["application/json"],"tags":["Wallet"],"summary":"Get wallet balance by user ID","parameters":[{"type":"integer","description":"User ID","name":"user_id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/wallet/create":{"post":{"security":[{"BearerAuth":[]}],"description":"Create a new wallet for a user","consumes":["application/json"],"produces":["application/json"],"tags":["Wallet"],"summary":"Create wallet","parameters":[{"description":"Wallet request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/models.Wallet"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/wallet/top-up":{"post":{"security":[{"BearerAuth":[]}],"description":"Add balance to a wallet","consumes":["application/json"],"produces":["application/json"],"tags":["Wallet"],"summary":"Top up wallet balance","parameters":[{"description":"Top up request","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/models.Wallet"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/wallet/user/{user_id}":{"get":{"security":[{"BearerAuth":[]}],"description":"Get wallet list by user ID","consumes":["application/json"],"produces":["application/json"],"tags":["Wallet"],"summary":"Get wallet by user ID","parameters":[{"type":"integer","description":"User ID","name":"user_id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}},"/api/v1/wallet/{id}":{"get":{"security":[{"BearerAuth":[]}],"description":"Get wallet details by wallet ID","consumes":["application/json"],"produces":["application/json"],"tags":["Wallet"],"summary":"Get wallet by ID","parameters":[{"type":"integer","description":"Wallet ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/model.Response"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/model.Response"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/model.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/model.Response"}}}}}},"definitions":{"model.Response":{"type":"object","properties":{"data":{},"error":{},"errors":{},"message":{"type":"string"},"messages":{"type":"array","items":{"type":"string"}},"page":{"type":"integer"},"per_page":{"type":"integer"},"status_code":{"type":"integer"},"total":{"type":"integer"}}},"models.InvoiceRequest":{"type":"object","properties":{"admin_id":{"type":"integer"},"amount":{"type":"number"},"buyer_id":{"type":"integer"},"created_at":{"type":"string"},"deleted_at":{"type":"string"},"invoice_id":{"type":"integer"},"invoice_name":{"type":"string"},"merchant_id":{"type":"integer"},"payment_type":{"type":"string"},"role_id":{"type":"string"},"status":{"type":"string"},"updated_at":{"type":"string"},"va_wallet":{"type":"number"}}},"models.Request":{"type":"object","properties":{"client_id":{"type":"string"},"client_secret":{"type":"string"},"code":{"type":"string"},"password":{"type":"string"},"redirect_uri":{"type":"string"},"user_token":{"type":"string"},"username":{"type":"string"}}},"models.Wallet":{"type":"object","properties":{"action":{"type":"string"},"created_at":{"type":"string"},"current_saldo":{"type":"number"},"deleted_at":{"type":"string"},"id":{"type":"integer"},"invoice_id":{"type":"integer"},"record_saldo":{"type":"number"},"status":{"type":"integer"},"updated_at":{"type":"string"},"user_id":{"type":"integer"}}}},"securityDefinitions":{"BasicAuth":{"type":"basic"},"BearerAuth":{"type":"apiKey","name":"Authorization","in":"header"}}}